Replay: Send indexed generation & serial numbers, not raw data.
To allow "fixing" a bad generation, we can manually update a bad generation
number or their serial numbers in the collector's database. The update only
applies to the generation_id and serial_no columns, not the raw_data which
stays unmodified. Therefore, to have any effect, we send the identifiers
found in the two columns, not the raw data.

This is really a hack for badly-behaving devices. We should really have a
generic tool to re-write the collector's database.